client/firmware_updater/requirements.txt
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
pycparser==2.23
PyYAML==6.0.3
requests==2.32.5
urllib3==2.5.0
